Refactoring for refactoring's sake. I wasn't happy with how things were split up
