+++ src/main/java/com/minelittlepony/model/ModelMobPony.java
@@ -14,13 +14,6 @@ public class ModelMobPony extends ModelAlicorn {
         super(false);
     }
 
-    @Override
-    protected void adjustLegs(float move, float swing, float tick) {
-        super.adjustLegs(move, swing, tick);
-        rotateRightArm(move, tick);
-        rotateLeftArm(move, tick);
-    }
-
     /**
      * Returns true if the angle is to the right?
      */
@@ -28,38 +21,25 @@ public class ModelMobPony extends ModelAlicorn {
         return MathHelper.sin(move / 20f) < 0;
     }
 
-    /**
-     * Called to update the left arm's final rotation.
-     * Subclasses may replace it with their own implementations.
-     *
-     * @param move Limb swing amount.
-     * @param tick Render partial ticks.
-     */
-    protected void rotateRightArm(float move, float tick) {
-        if (rightArmPose == ArmPose.EMPTY) return;
-
-        if (canCast()) {
-            unicornArmRight.setRotationPoint(-7, 12, -2);
-            rotateArmHolding(unicornArmRight, -1, swingProgress, tick);
-        } else {
-            rotateArmHolding(bipedRightArm, -1, swingProgress, tick);
+    @Override
+    protected void adjustLegs(float move, float swing, float tick) {
+        super.adjustLegs(move, swing, tick);
+        if (rightArmPose != ArmPose.EMPTY) {
+            if (canCast()) {
+                unicornArmRight.setRotationPoint(-7, 12, -2);
+                rotateArmHolding(unicornArmRight, -1, swingProgress, tick);
+            } else {
+                rotateArmHolding(bipedRightArm, -1, swingProgress, tick);
+            }
         }
-    }
 
-    /**
-     * Same as rotateRightArm but for the left arm (duh).
-     *
-     * @param move Limb swing amount.
-     * @param tick Render partial ticks.
-     */
-    protected void rotateLeftArm(float move, float tick) {
-        if (leftArmPose == ArmPose.EMPTY) return;
-
-        if (!canCast()) {
-            unicornArmRight.setRotationPoint(-7, 12, -2);
-            rotateArmHolding(unicornArmLeft, -1, swingProgress, tick);
-        } else {
-            rotateArmHolding(bipedLeftArm, -1, swingProgress, tick);
+        if (leftArmPose != ArmPose.EMPTY) {
+            if (!canCast()) {
+                unicornArmRight.setRotationPoint(-7, 12, -2);
+                rotateArmHolding(unicornArmLeft, -1, swingProgress, tick);
+            } else {
+                rotateArmHolding(bipedLeftArm, -1, swingProgress, tick);
+            }
         }
     }
 }
