+++ src/main/java/com/minelittlepony/model/armour/ModelZombiePonyArmor.java
@@ -3,10 +3,10 @@ package com.minelittlepony.model.armour;
 import com.minelittlepony.render.AbstractPonyRenderer;
 
 public class ModelZombiePonyArmor extends ModelPonyArmor {
-
     // Copied from ModelZombiePony
     @Override
-    protected void rotateRightArm(float move, float tick) {
+    protected void adjustLegs(float move, float swing, float tick) {
+        super.adjustLegs(move, swing, tick);
         if (rightArmPose != ArmPose.EMPTY) return;
 
         if (islookAngleRight(move)) {
@@ -14,11 +14,6 @@ public class ModelZombiePonyArmor extends ModelPonyArmor {
         } else {
             rotateArmHolding(bipedLeftArm, -1, swingProgress, tick);
         }
-    }
-
-    @Override
-    protected void rotateLeftArm(float move, float tick) {
-        // Zombies are unidexterous.
     }
 
     @Override
